mbta-bus version 4
This switches mbta-bus to a Vue.js app, and uses the official MBTA v3
JSON-API, since version 2 has been discontinued.

It fixes a lot of bugs and is more reactive, and a nice intro to
modern Javascript and Vue.js for me.  All the features of the old site
should continue to work.

It drops the KML stuff entirely, as the v3 API provides me with
polyline paths for routes that I can parse and add directly.  It also
means we don't need the GTFS data.